Added the MatchingBracket class and changed the model to find existing pairs

+++ Hacking/PuzzleWord.cpp
@@ -16,13 +16,13 @@ void PuzzleWord::SetPosition( int position, int columnCount, int columnHeight, i
 {
     this->position = position;
 
-    for ( int i = 0; i < this->letterPositions.size(); ++i )
+    for ( int i = 0; i < ( int )this->letterPositions.size(); ++i )
     {
         LetterPosition& letterPos = this->letterPositions[i];
 
         letterPos.position = position + i;
         letterPos.column = letterPos.position / ( columnHeight * columnWidth );
-        letterPos.x = (letterPos.position % ( columnHeight * columnWidth )) % columnWidth;
+        letterPos.x = ( letterPos.position % ( columnHeight * columnWidth ) ) % columnWidth;
         assert( letterPos.x >= 0 && letterPos.x < columnWidth );
         letterPos.y = ( letterPos.position % ( columnHeight * columnWidth ) ) / columnWidth;
         assert( letterPos.y >= 0 && letterPos.y < columnHeight );
@@ -31,12 +31,12 @@ void PuzzleWord::SetPosition( int position, int columnCount, int columnHeight, i
 
 const LetterPosition & PuzzleWord::GetLetterPosition( int index ) const
 {
-    assert( index >= 0 && index < this->text.size() );
+    assert( index >= 0 && index < ( int )this->text.size() );
     return this->letterPositions[index];
 }
 
 LetterPosition & PuzzleWord::GetLetterPosition( int index )
 {
-    assert( index >= 0 && index < this->text.size() );
+    assert( index >= 0 && index < ( int )this->text.size() );
     return this->letterPositions[index];
 }
